requirements.txt
torch==1.8.1
numpy==1.19.5
ipython==7.28.0
PyYAML==5.4.1
